feat: 히스토리 api 수정중

+++ RunMate/domain/src/main/java/com/D107/runmate/domain/model/history/GroupRun.kt
@@ -3,10 +3,10 @@ package com.D107.runmate.domain.model.history
 import com.D107.runmate.domain.model.base.BaseModel
 
 data class GroupRun(
-    val avgPace: Int,
+    val avgPace: Double,
     val courseLiked: Boolean,
-    val distance: Int,
+    val distance: Double,
     val nickname: String,
-    val time: Int,
+    val time: Double,
     val userId: String
 ): BaseModel
